refactor: extract the request logic to a separate helper
